FROMLIST: xhci: workaround for hosts missing CAS bit
If a device is unplugged and replugged during Sx system suspend
some  Intel xHC hosts will overwrite the CAS (Cold attach status) flag
and no device connection is noticed in resume.

A device in this state can be identified in resume if its link state
is in polling or compliance mode, and the current connect status is 0.
A device in this state needs to be warm reset.

Intel 100/c230 series PCH specification update Doc #332692-006 Errata #8

Observed on Cherryview and Apollolake as they go into compliance mode
if LFPS times out during polling, and re-plugged devices are not
discovered at resume.

Signed-off-by: Mathias Nyman <mathias.nyman@linux.intel.com>
CC: <stable@vger.kernel.org

Link: //http://marc.info/?l=linux-usb&m=147697618721027&w=2
(Removed APL reference from original patch)

BUG=chrome-os-partner:52727
TEST=Connect usb3 device, enter s3, hotplug the usb3 device, resume
from s3 and check if the usb3 device is still detected.

Change-Id: Ic4419c7271133455b825aeb0a3c41b2a3100211b
Signed-off-by: Jagadish Krishnamoorthy <jagadish.krishnamoorthy@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/401404
Tested-by: Robert Huang <robert.huang@quantatw.com>
Tested-by: Huki Huang <huki.huang@intel.com>
Reviewed-by: Benson Leung <bleung@chromium.org>