CHROMIUM: LSM: Add support for symlink traversal policy.
This change adds functionality to the chromiumos LSM to pin inodes and
associate tracking information with them. This allows tracking symlink
traversal policy for an inode. Whenever the FS layer follows a link,
it calls out to the LSM, which can then check whether a symlink
traversal policy has been set on any directory on the path in
question. If a policy is set and it specifies that symlink traversal
should be blocked, the corresponding FS system call will fail with
EPERM. Note that readlink() continues to work, so userspace may resort
to resolving symlinks manually if desired.

Symlink traversal policy is configured via securityfs. Files in
/sys/kernel/security/chromiumos/symlink_policy/ are used to set the
symlink traversal policy:
  * .../allow: Write a path to the file to set the symlink traversal
    policy for corresponding inode to "allow".
  * .../block: Write a path to the file to set the symlink traversal
    policy for corresponding inode to "block".
  * .../reset: Write a path to clear the symlink traversal policy for
    the corresponding inode.
  * .../flush: Write a path to clear symlink traversal policy for the
    entire file system the corresponding inode belongs to. Note that
    this may affect paths outside the subtree corresponding to the
    specified path (e.g. parents, siblings, or parts of the file
    system not visible in the affected (bind) mount).

BUG=chromium:655606
TEST=Set symlink traversal policy and test paths containing symlinks.

Change-Id: Ia31fe20d18ba31857a96a6c846aeca767328a018
Reviewed-on: https://chromium-review.googlesource.com/422334
Commit-Ready: Mattias Nissler <mnissler@chromium.org>
Tested-by: Mattias Nissler <mnissler@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>