Include I2C_CHARDEV for x86 on 3.18
With this CL we change the status of I2C_CHARDEV to compiling it
directly into the kernel for x86 devices on kernel 3.18.  This
driver is required for some of our devices to update a boot and
if it's a module it sometimes is not availible early enough.

BUG=chrome-os-partner:59953
TEST=built the kernel and deployed on the Chromebook, everything
worked well at boot whereas before it was failing -- the FW
updater that needed it worked smoothly with this kernel.

Change-Id: I3963db4f24774d80a7b6ec97cfe474a03a9cd44f
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/412123