CHROMIUM: drm/img-rogue: Fix firmware initialisation on device rebind
A local static variable was used in SrvInit() to prevent the firmware from
being reinitialised each time the device was opened. However, this also
prevented the device from being successfully rebound to the driver.

Remove the use of the local static variable by instead using the init server
state to determine whether or not the firmware has already been initialised.
Reset this state when the device is unbound from the driver.

BUG=chrome-os-partner:58144
TEST=stop ui
     echo 13000000.mfgsys-gpu > /sys/bus/platform/drivers/pvrsrvkm_1_6/unbind
	 echo 13000000.mfgsys-gpu > /sys/bus/platform/drivers/pvrsrvkm_1_6/bind
	 start ui

Change-Id: I68d7aaf3961343826757834f28ae887687a8e1a0
Signed-off-by: Frank Binns <frank.binns@imgtec.com>
Reviewed-on: https://chromium-review.googlesource.com/413266
Commit-Ready: Nicolas Boichat <drinkcat@chromium.org>
Tested-by: Nicolas Boichat <drinkcat@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>