CHROMIUM: config: turn off CMA on arm64
There's no known reason why we have CMA turned on for arm64.
It was enabled during early bring-up of a platform on chromeos 3.18, and
was never turned off.
For all arm64 configs in kernel 4.4 we've got iommus and those are used
when we want a contiguous chunk of memory for graphics or video.

Let's get rid of a bunch of extra code and complexity in the kernel by
turning off CMA.

Inspired by the chromeos-4.4 equivalent patch:
https://chromium-review.googlesource.com/414433

Signed-off-by: Daniel Kurtz <djkurtz@chromium.org>

BUG=chrome-os-partner:60243
TEST=for i in $(seq 100); do ./launchBalloons.sh 6 700 30 >/dev/null; done
TEST=grep cma /var/log/message
 => no line like cma: Reserved 16 MiB at 0x00000000fec00000
Change-Id: I637475f947323af0c4fbae02f7dd4b5fb8e817e2
Reviewed-on: https://chromium-review.googlesource.com/416839
Commit-Ready: Daniel Kurtz <djkurtz@chromium.org>
Tested-by: Daniel Kurtz <djkurtz@chromium.org>
Reviewed-by: Nicolas Boichat <drinkcat@chromium.org>