CHROMIUM: drm/udl: flush physical pages after allocating
udl_gem_get_pages gets pages which are potentially in the CPU
cache with stale contents. This memory incoherency causes
corruption which is visible on an external monitor. To fix this,
flush pages that are fetched from the cache.

This is not an issue for systems that address the cache using virtual
memory bits.

BUG=chromium:668636
TEST=trybots

Change-Id: Iba90c5be70b6d32982ec7f8822f1e5658d25fbad
Reviewed-on: https://chromium-review.googlesource.com/418009
Commit-Ready: Gurchetan Singh <gurchetansingh@chromium.org>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>