FROMLIST: vb2: Move cache synchronisation from buffer done to dqbuf handler
The cache synchronisation may be a time consuming operation and thus not
best performed in an interrupt which is a typical context for
vb2_buffer_done() calls. This may consume up to tens of ms on some
machines, depending on the buffer size.

Signed-off-by: Sakari Ailus <sakari.ailus@linux.intel.com>

(am from https://patchwork.kernel.org/patch/9477165/)

BUG=chrome-os-partner:43703
TEST=Compiled and tested on Elm.

Change-Id: I6b454f461ad57d0fa80c32aa4ec45d71e881435a
Signed-off-by: Ricky Liang <jcliang@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/422719
Reviewed-by: Tomasz Figa <tfiga@chromium.org>