CHROMIUM: Platform: x86 : chromeos_laptop - Add caroline support
Caroline has a PEN_EJECT signal that is routed to 2 CPU GPIO pins:

B15: This is used for wakeup (configured by ACPI)
B19: We need an input driver to use this to emit SW_PEN_INSERTED event.
     (This patch enables the B19 usecase).

Kevin uses gpio-keys driver to do this but the driver only understands
device trees or platform data. Hence we instantiate the gpio-keys
with the platform data.

It tries to find the GPIO by finding the right Intel gpio_chip and
pin 43 (B19) under that chip.

I contemplated doing this exactly like it is done for I2C devices currently
(thought of introducing a array of platform_devices in the chromeos_laptop
structure and instantiating them in chromeos_laptop_probe()). But since
I wanted to search for gpio chip, hence I added a platform_init() hook.

BUG=chrome-os-partner:59763
TEST=Verify that the "stylus tools" menu gets launched when I eject
the pen on caroline.

Change-Id: Iac645accb0a6687640c7ef48487f99ae4cf44bb5
Reviewed-on: https://chromium-review.googlesource.com/424941
Commit-Ready: Rajat Jain <rajatja@chromium.org>
Tested-by: Rajat Jain <rajatja@chromium.org>
Reviewed-by: Benson Leung <bleung@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>